refactor(game): move ttf_SDL2 declaration into separate header
