+++ src/game/graphics/SDL2/graphics_SDL2.h
@@ -16,7 +16,6 @@ struct SDL_RWops;
 struct SDL_Renderer;
 struct SDL_Surface;
 struct SDL_Texture;
-using TTF_Font = struct _TTF_Font; // NOLINT(cert-dcl37-c,cert-dcl51-cpp,bugprone-reserved-identifier): SDL detail
 
 // global control pointer, do not modify
 inline SDL_Renderer* gFrameRenderer;
@@ -111,23 +110,4 @@ private:
 public:
     TextureFull(const Color& srcColor);
     ~TextureFull() override;
-};
-
-////////////////////////////////////////////////////////////////////////////////
-// SDL_ttf encapsulation. Mostly as same as Image
-// Run TTF_Init outside.
-class TTFFont
-{
-private:
-    TTF_Font* _pFont = nullptr;
-    std::string _filePath;
-    bool loaded = false;
-
-public:
-    TTFFont(const Path& filePath, int ptsize);
-    TTFFont(const Path& filePath, int ptsize, int faceIndex);
-    ~TTFFont();
-
-    std::shared_ptr<Texture> TextUTF8(const char* text, const Color& c);
-    [[nodiscard]] bool isLoaded() const { return loaded; }
 };
